Renamed common/utils/enum.js to common/utils/enums.js
