+++ frontend/jsx/common/utils/enums.js
